@@ -1,7 +1,19 @@
-import React, {Component} from 'react';
+import React, {useState} from 'react'
 
-export default class Todo extends Component {
-  render() {
-    return <p>{this.props.task}</p>;
-  }
+const Todo = (props) => {
+    return (
+        <div>
+            {props.task}
+        </div>
+    )
 }
+
+export default Todo
+
+// import React, {Component} from 'react';
+
+// export default class Todo extends Component {
+//   render() {
+//     return <p>{this.props.task}</p>;
+//   }
+// }
